Optimize multiple-instance deployment per container
We are really not using a lot of RAM, so it'd be
rather wasteful to be using many cloud service
instances with 8-16GB just for the same of accessing
their vCPUs to leverage multithreading.

Instead, we'll deploy multiple instances of the
worker service to a given container, allowing for
less waste on RAM and a little bit more vertical
scaling.